@@ -10,21 +10,23 @@
   },
   "dependencies": {
     "@next-auth/prisma-adapter": "^1.0.7",
-    "@types/node": "20.5.7",
-    "@types/react": "18.2.21",
-    "@types/react-dom": "18.2.7",
-    "autoprefixer": "10.4.15",
-    "eslint": "8.48.0",
-    "eslint-config-next": "13.4.19",
+    "@types/node": "20.8.7",
+    "@types/react": "18.2.31",
+    "@types/react-dom": "18.2.14",
+    "autoprefixer": "10.4.16",
+    "eslint": "8.52.0",
     "faunadb": "^4.8.0",
-    "next": "13.4.19",
-    "next-auth": "^4.23.2",
+    "next": "^13.5.6",
+    "next-auth": "^4.24.3",
     "nodemailer": "^6.9.6",
-    "postcss": "8.4.29",
+    "postcss": "8.4.31",
     "prisma": "^5.4.2",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "tailwindcss": "3.3.3",
     "typescript": "5.2.2"
+  },
+  "devDependencies": {
+    "eslint-config-next": "^13.5.6"
   }
 }
