@@ -1,13 +1,16 @@
+'use client'
+
 import React from "react";
 import { useSession } from "next-auth/react";
-import Router from 'next/router';
+import { useRouter } from 'next/navigation'
 
 const Connections = () => {
 
     const { status } = useSession()
+    const router = useRouter()
 
     if (status === "unauthenticated") {
-        Router.push('/api/auth/signin')
+        router.push('/api/auth/signin')
     }
 
     return <div>Connections</div>;
